Correction migration

+++ database/migrations/2021_06_30_091033_create_commandes_table.php
@@ -17,7 +17,7 @@ class CreateCommandesTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id')->index();
             $table->string('mode_livraison');
-            $table->integer('prix');
+            $table->float('prix');
             $table->timestamps();
 
             //Foreign keys
